parser accepts domain annotation, gen possibility via constraints
In order to avoid issues with safety any predicate which takes variales should be annotated with the domain of the variable.

The bison parser has been updated with an inital parsing of this annotation.
For now this is the same as conditions (a colon).

The program built to generate the complement/converse has been updated to include these annotations.
To do this, the mapping has been revised to NLiterals in order to access annotations.

There's a lot of rushed code at the moment, and if this goes further than proof of concept it should be cleaned up.