Examples used for evaluating bounded FastLAS
In short, generated by combining the chunking examples from FastLAS1.
Bounded examples correspond to each unique context with added true_split predicates generated by taking the union of inclusions for each distinct context.
The bound for each example is equal to the number of true splits.

Examples are subdivided according to names in the FastLAS1 folder.

The unified examples folder collects together all examples for a dataset in a single file.
The separate_bes folder stores each example in a different file together with a (common) background file to make evaluating easier.